seperate logic into builders
