+++ server/utils/filterBuilder.ts
@@ -1,0 +1,15 @@
+export default function filterBuilder(filter: object) {
+  let params: string[] = [];
+  let values: string[] = [];
+  // Add filters dynamically to query and values array
+  Object.entries(filter).forEach(([key, value]) => {
+    const endsWith =
+      value.startsWith("%") || value.endsWith("%") ? " LIKE" : " =";
+    params.push(`${key}${endsWith} ?`);
+    console.log(value);
+
+    values.push(value as string);
+  });
+
+  return { params, values };
+}
